+++ classes/class.ilSrUserEnrolmentUIHookGUI.php
@@ -1,6 +1,8 @@
 <?php
 
 use srag\DIC\SrUserEnrolment\DICTrait;
+use srag\Plugins\SrUserEnrolment\Config\Config;
+use srag\Plugins\SrUserEnrolment\ResetPassword\ResetPasswordGUI;
 use srag\Plugins\SrUserEnrolment\Rule\Repository;
 use srag\Plugins\SrUserEnrolment\Rule\RulesGUI;
 use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
@@ -16,6 +18,8 @@ class ilSrUserEnrolmentUIHookGUI extends ilUIHookPluginGUI {
 	use SrUserEnrolmentTrait;
 	const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
 	const PAR_SUB_TABS = "sub_tabs";
+	const COURSE_MEMBER_LIST_TEMPLATE_ID = "Services/Table/tpl.table2.html";
+	const TEMPLATE_GET = "template_get";
 
 
 	/**
@@ -23,6 +27,63 @@ class ilSrUserEnrolmentUIHookGUI extends ilUIHookPluginGUI {
 	 */
 	public function __construct() {
 
+	}
+
+
+	/**
+	 * @param string $a_comp
+	 * @param string $a_part
+	 * @param array  $a_par
+	 *
+	 * @return array
+	 */
+	public function getHTML(/*string*/ $a_comp, /*string*/ $a_part, $a_par = []): array {
+		if (self::dic()->ctrl()->getCmdClass() === strtolower(ilCourseMembershipGUI::class)
+			&& (empty(self::dic()->ctrl()->getCmd())
+				|| self::dic()->ctrl()->getCmd() === "participants")) {
+
+			if ($a_par["tpl_id"] === self::COURSE_MEMBER_LIST_TEMPLATE_ID && $a_part === self::TEMPLATE_GET) {
+
+				if (Config::getField(Config::KEY_SHOW_RESET_PASSWORD)) {
+
+					if (self::access()->currentUserHasRole()) {
+
+						$html = $a_par["html"];
+
+						$course = new ilObjCourse(self::rules()->getRefId());
+
+						$html = preg_replace_callback('/<a class="il_ContainerItemCommand2" href=".+member_id=([0-9]+).+cmd=editMember.+">.+<\/a>/', function (array $matches) use ($course): string {
+							$link = $matches[0];
+
+							$user_id = intval($matches[1]);
+
+							if (self::ilias()->courses()->isMember($course, $user_id)) {
+								self::dic()->ctrl()->saveParameterByClass(ResetPasswordGUI::class, Repository::GET_PARAM_REF_ID);
+
+								self::dic()->ctrl()->setParameterByClass(ResetPasswordGUI::class, Repository::GET_PARAM_USER_ID, $user_id);
+
+								return self::output()->getHTML([
+									$link,
+									"<br>",
+									self::dic()->ui()->factory()->link()->standard(self::plugin()
+										->translate("button", ResetPasswordGUI::LANG_MODULE_RESET_PASSWORD), self::dic()->ctrl()
+										->getLinkTargetByClass([
+											ilUIPluginRouterGUI::class,
+											ResetPasswordGUI::class
+										], ResetPasswordGUI::CMD_RESET_PASSWORD_CONFIRM))
+								]);
+							} else {
+								return $link;
+							}
+						}, $html);
+
+						return [ "mode" => self::REPLACE, "html" => $html ];
+					}
+				}
+			}
+		}
+
+		return parent::getHTML($a_comp, $a_part, $a_par);
 	}