+++ src/Access/Courses.php
@@ -2,18 +2,19 @@
 
 namespace srag\Plugins\SrUserEnrolment\Access;
 
+use ilObjCourse;
 use ilSrUserEnrolmentPlugin;
 use srag\DIC\SrUserEnrolment\DICTrait;
 use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
 
 /**
- * Class Ilias
+ * Class Courses
  *
  * @package srag\Plugins\SrUserEnrolment\Access
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */
-final class Ilias {
+final class Courses {
 
 	use DICTrait;
 	use SrUserEnrolmentTrait;
@@ -37,7 +38,7 @@ final class Ilias {
 
 
 	/**
-	 * Ilias constructor
+	 * Courses constructor
 	 */
 	private function __construct() {
 
@@ -45,33 +46,12 @@ final class Ilias {
 
 
 	/**
-	 * @return Courses
+	 * @param ilObjCourse $course
+	 * @param int         $user_id
+	 *
+	 * @return bool
 	 */
-	public function courses(): Courses {
-		return Courses::getInstance();
-	}
-
-
-	/**
-	 * @return OrgUnits
-	 */
-	public function orgUnits(): OrgUnits {
-		return OrgUnits::getInstance();
-	}
-
-
-	/**
-	 * @return Roles
-	 */
-	public function roles(): Roles {
-		return Roles::getInstance();
-	}
-
-
-	/**
-	 * @return Users
-	 */
-	public function users(): Users {
-		return Users::getInstance();
+	public function isMember(ilObjCourse $course, int $user_id): bool {
+		return $course->getMembersObject()->isMember($user_id);
 	}
 }