+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## [x]
+- Reset password from course member list
+
 ## [1.0.0]
 - First version